Update Utility_working.py
Included another criteria for the Response counting in which if the Start and End event times BOTH fall outside the range of the window then it still counts as an event. Previously at least one Start or End had to fall within the window, but some event windows are very small and some ongoing activity triggered a 'start' before the actual event.