Event counter fix
In event counter, the variable S needs to be initialized as 0 not 1 as you had initially written. If it starts at 1, then the search for a stim window begins at the 2nd row rather than the 1st row which was and when S was used to lookup the correct threshold, it was shifted by 1 because the thresholds have a title row, whereas the Stim array does not.